docs: README 화면 미리보기 이미지 추가
